os-hubline/backend/requirements.txt
# OS HubLine — Backend Dependencies
# Framework
fastapi==0.115.0
uvicorn[standard]==0.30.0
pydantic==2.9.0
pydantic-settings==2.5.0

# Database — Core
sqlalchemy[asyncio]==2.0.35
alembic==1.13.0

# Database — Drivers (installer celui correspondant à votre moteur)
# PostgreSQL (défaut)
asyncpg==0.29.0
psycopg2-binary==2.9.9
# SQL Server
# aioodbc==0.5.0
# pyodbc==5.1.0
# MySQL / MariaDB
# aiomysql==0.2.0
# pymysql==1.1.0
# SQLite (inclus dans Python, driver async ci-dessous)
aiosqlite==0.20.0
# Oracle
# oracledb==2.4.0

# Auth & Security
pyjwt==2.9.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.9

# Task Queue
celery[redis]==5.4.0
redis==5.1.0

# HTTP Client (integrations)
httpx==0.27.0

# Validation
email-validator==2.2.0

# Monitoring
prometheus-client==0.21.0
sentry-sdk[fastapi]==2.14.0

# Testing
pytest==8.3.0
pytest-asyncio==0.24.0
httpx==0.27.0
factory-boy==3.3.0

# Utils
python-dotenv==1.0.1
orjson==3.10.0
